Updated site registration guards
