+++ prime-angular-frontend/src/app/modules/site-registration/shared/guards/organization.guard.ts
@@ -11,38 +11,176 @@ import { OrganizationResource } from '@core/resources/organization-resource.serv
 
 import { AuthService } from '@auth/shared/services/auth.service';
 
+import { SiteRoutes } from '@registration/site-registration.routes';
 import { Organization } from '@registration/shared/models/organization.model';
 import { OrganizationService } from '@registration/shared/services/organization.service';
 
-// TODO PRIME-1131
-// Should this guard be dropped and only use the registration guard
 @Injectable({
   providedIn: 'root'
 })
 export class OrganizationGuard extends BaseGuard {
+
   constructor(
     protected authService: AuthService,
     protected logger: LoggerService,
     @Inject(APP_CONFIG) private config: AppConfig,
     private router: Router,
-    private organizationResource: OrganizationResource,
-    private organizationService: OrganizationService
+    private organizationService: OrganizationService,
+    private organizationResource: OrganizationResource
   ) {
     super(authService, logger);
   }
 
   protected checkAccess(routePath: string = null, params: Params): Observable<boolean> | Promise<boolean> {
-    const organizationId = params.oid;
-    return this.organizationResource.getOrganizationById(organizationId)
+    return this.organizationResource.getOrganizations()
       .pipe(
+        map((organizations: Organization[]) =>
+          (organizations.length)
+            ? organizations.shift()
+            : null
+        ),
         map((organization: Organization) => {
-          // Store the organization for access throughout creation and updating of a
-          // organization, which will allows provide the most up-to-date organization
+          // Store the organization for access throughout registration, which
+          // will allows be the most up-to-date organization
           this.organizationService.organization = organization;
-          // TODO PRIME-1131
-          // TODO always resolve until routes are lock down
-          return true;
+          // Determine the next route based on whether this is the initial
+          // registration of an organization and site, or subsequent
+          // registration of sites under an existing organization
+          return this.routeDestination(routePath, organization);
         })
       );
+  }
+
+  /**
+   * @description
+   * Determine the route destination based on the organization status.
+   */
+  private routeDestination(routePath: string, organization: Organization) {
+    // On login the user will always be redirected to the collection notice
+    if (routePath.includes(SiteRoutes.COLLECTION_NOTICE)) {
+      return true;
+    } else if (organization) {
+      return (organization.completed)
+        ? this.manageCompleteOrganizationRouting(routePath, organization)
+        : this.manageIncompleteOrganizationRouting(routePath, organization);
+    }
+
+    // Otherwise, prevent the route from resolving
+    return false;
+  }
+
+  /**
+   * @description
+   * Manage routing when an organization exists.
+   */
+  private manageCompleteOrganizationRouting(routePath: string, organization: Organization) {
+    // Provides a default of the main management view unless current view
+    // can be determined through state of the organization
+    return this.manageRouting(routePath, SiteRoutes.SITE_MANAGEMENT, organization);
+  }
+
+  /**
+   * @description
+   * Manage routing when an organization does not exist, or
+   * initial registration has not been completed.
+   */
+  private manageIncompleteOrganizationRouting(routePath: string, organization: Organization) {
+    // Provides a default of the initial site registration view unless the current view
+    // can be determined through state of the organization
+    return this.manageRouting(routePath, SiteRoutes.ORGANIZATION_SIGNING_AUTHORITY, organization);
+  }
+
+  private manageRouting(routePath: string, defaultRoute: string, organization: Organization): boolean {
+    const currentRoute = this.getCurrentRoute(routePath);
+    const allowedRoutes = this.getAllowedRoutes(organization);
+
+    // Redirect to an appropriate default route when the current
+    // route does not exist in the list of allowed routes
+    if (!allowedRoutes.includes(currentRoute)) {
+      return (organization.completed)
+        ? this.navigate(routePath, SiteRoutes.SITE_MANAGEMENT)
+        : this.navigate(routePath, SiteRoutes.SITE_MANAGEMENT, defaultRoute, organization.id);
+    }
+
+    // Otherwise, allow access to the route
+    return true;
+  }
+
+  /**
+   * @description
+   * Get the route URI param that is the current route attempting
+   * to be resolved, or replace as required.
+   */
+  private getCurrentRoute(routePath: string) {
+    // Remote users has a child view that should not be a point of
+    // redirection so the parent is purposefully targeted, otherwise
+    // get the last URI param to determine the current route
+    let currentRoute = routePath.includes(SiteRoutes.REMOTE_USERS)
+      ? SiteRoutes.REMOTE_USERS
+      : routePath.split('/').pop();
+
+    // Strip off queries when they exist
+    if (currentRoute.includes('?')) {
+      currentRoute = currentRoute.split('?')[0];
+    }
+
+    return currentRoute;
+  }
+
+  /**
+   * @description
+   * Get a list of allowed routes based on the state of the creation
+   * of an organization during an initial registration, otherwise
+   * attempt to redirect to an appropriate point during a registration.
+   */
+  private getAllowedRoutes(organization: Organization) {
+    // Allow access to a set of routes based on initial registration
+    // having been completed (default)
+    let allowedRoutes = SiteRoutes.siteRegistrationRoutes();
+
+    if (!organization.completed) {
+      // Initial organization creation has not been completed and should
+      // be replaced with the allowed initial registration route order
+      allowedRoutes = allowedRoutes
+        .filter((route: string) =>
+          SiteRoutes.organizationRegistrationRouteOrder().includes(route)
+        );
+    } else {
+      // Completed indicates the data entry has occurred, but does not
+      // mean the organization agreement has been accepted or the
+      // registration submitted so allow access to these routes
+      allowedRoutes.push(SiteRoutes.ORGANIZATION_AGREEMENT);
+      allowedRoutes.push(SiteRoutes.ORGANIZATION_REVIEW);
+
+      if (organization.hasAcceptedAgreement) {
+        allowedRoutes.push(SiteRoutes.SITE_REVIEW);
+      }
+    }
+
+    return allowedRoutes;
+  }
+
+  /**
+   * @description
+   * Prevent infinite route loops by navigating to a route only
+   * when the current route path is not the destination path.
+   */
+  private navigate(
+    routePath: string,
+    loopPath: string,
+    destinationPath: string = null,
+    oid: number = null
+  ): boolean {
+    const modulePath = this.config.routes.site;
+    const comparePath = (destinationPath && oid)
+      ? `/${modulePath}/${loopPath}/${oid}/${destinationPath}`
+      : `/${modulePath}/${loopPath}`;
+
+    if (routePath === comparePath) {
+      return true;
+    } else {
+      this.router.navigate([comparePath]);
+      return false;
+    }
   }
 }
