--- prime-angular-frontend/src/app/modules/site-registration/shared/guards/registration.guard.spec.ts
@@ -1,36 +1,0 @@
-import { TestBed, waitForAsync, inject } from '@angular/core/testing';
-
-import { RegistrationGuard } from './registration.guard';
-import { RouterTestingModule } from '@angular/router/testing';
-import { APP_CONFIG, APP_DI_CONFIG } from 'app/app-config.module';
-import { AuthService } from '@auth/shared/services/auth.service';
-import { MockAuthService } from 'test/mocks/mock-auth.service';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { SharedModule } from '@shared/shared.module';
-
-describe('RegistrationGuard', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        RouterTestingModule,
-        HttpClientTestingModule,
-        SharedModule
-      ],
-      providers: [
-        RegistrationGuard,
-        {
-          provide: APP_CONFIG,
-          useValue: APP_DI_CONFIG
-        },
-        {
-          provide: AuthService,
-          useClass: MockAuthService
-        }
-      ]
-    });
-  });
-
-  it('should ...', inject([RegistrationGuard], (guard: RegistrationGuard) => {
-    expect(guard).toBeTruthy();
-  }));
-});
